requirements.txt
streamlit==1.31.0
opencv-python==4.9.0.80
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
Pillow==10.2.0
